+++ src/main/kotlin/com/flyn/sarcopenia_server/data/RawMessage.kt
@@ -1,0 +1,5 @@
+package com.flyn.sarcopenia_server.data
+
+import io.netty.buffer.ByteBuf
+
+class RawMessage(val code: Byte, val message: ByteBuf)