+++ src/main/kotlin/com/flyn/sarcopenia_server/data/ResponseData.kt
@@ -1,4 +1,0 @@
-package com.flyn.sarcopenia_server.data
-
-data class ResponseData(var value: Int) {
-}