+++ src/main/kotlin/com/flyn/sarcopenia_server/decoder/DecoderState.kt
@@ -1,0 +1,5 @@
+package com.flyn.sarcopenia_server.decoder
+
+enum class DecoderState {
+    MESSAGE_TYPE, MESSAGE_LENGTH, MESSAGE
+}