+++ src/main/kotlin/com/flyn/sarcopenia_server/handler/ConnectionHandler.kt
@@ -1,0 +1,12 @@
+package com.flyn.sarcopenia_server.handler
+
+import io.netty.channel.ChannelHandlerContext
+import io.netty.channel.ChannelInboundHandlerAdapter
+
+class ConnectionHandler: ChannelInboundHandlerAdapter() {
+
+    override fun channelActive(ctx: ChannelHandlerContext) {
+        println(ctx.channel().remoteAddress())
+    }
+
+}