+++ CHANGELOG.md
@@ -5,9 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Spilt message data to other project
+### Changed
+- Let channel control the user's uuid for reduce message size
 
 ## [0.1.0] - 2022-03-18
 ### Added
 - Basic server model
 - Request and response data
 - Data decoder and encoder
+
+## [0.2.0] - 2022-3-29
+### Added
+- Exchange file function
+- Commandline
+### Changed
+- Rebuild server construct
+### Removed
+- Request and response data